reworked UserInfoEdit screen
