+++ src/views/ModalViews/ApplicationForm.vue
@@ -3,7 +3,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Firma:
         </p>
@@ -13,7 +13,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Beruf:
         </p>
@@ -23,7 +23,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Deadline:
         </p>
