+++ src/views/ModalViews/EducationEdit.vue
@@ -3,7 +3,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Typ:
         </p>
@@ -17,7 +17,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Schwerpunkt:
         </p>
@@ -31,7 +31,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Adresse:
         </p>
@@ -45,7 +45,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Von:
         </p>
@@ -55,7 +55,7 @@
     <div class="col-span-2 md:col-span-1">
       <div class="flex bg-white dark:bg-slate-800 h-10">
         <p
-          class="px-1 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
+          class="px-0 py-2 w-32 h-10 text-black dark:text-white font-Montserrat text-base md:text-md"
         >
           Bis:
         </p>
